requirements.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
awsebcli==3.19.1
botocore==1.17.63
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.7
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.15.2
future==0.16.0
idna==2.7
iniconfig==1.0.1
isort==5.5.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
pathspec==0.5.9
pep8==1.7.1
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
pytest==6.1.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
urllib3==1.24.3
wcwidth==0.1.9
wrapt==1.12.1
